+++ src/pages/login/index.tsx
@@ -72,9 +72,5 @@ export const getServerSideProps: GetServerSideProps<object> = async (
 
 	return {
 		props: {},
-		redirect: {
-			destination: '/dashboard',
-			permanent: false,
-		},
 	};
 };
